feat(member) : role 필드 추가 및 저장/조회 로직 반영
