--- src/main/java/gift/dto/MemberRegisterRequest.java
@@ -1,4 +1,0 @@
-package gift.dto;
-
-public class MemberRegisterRequest {
-}
