+++ src/main/java/gift/repository/MemberRepositoryImpl.java
@@ -1,0 +1,41 @@
+package gift.repository;
+
+import gift.domain.Member;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public class MemberRepositoryImpl implements MemberRepository {
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public MemberRepositoryImpl(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
+    public Member save(Member member) {
+        String sql = "INSERT INTO members (email, password, role) VALUES (?, ?, ?)";
+        jdbcTemplate.update(sql, member.getEmail(), member.getPassword(), member.getRole());
+
+        return findByEmail(member.getEmail()).orElseThrow(); // 저장 후 다시 조회
+    }
+
+    @Override
+    public Optional<Member> findByEmail(String email) {
+        String sql = "SELECT * FROM members WHERE email = ?";
+        return jdbcTemplate.query(sql, memberRowMapper(), email).stream().findFirst();
+    }
+
+    private RowMapper<Member> memberRowMapper() {
+        return (rs, rowNum) -> new Member(
+                rs.getLong("id"),
+                rs.getString("email"),
+                rs.getString("password"),
+                rs.getString("role")
+        );
+    }
+}
