--- src/main/java/gift/service/MemberService.java
@@ -1,4 +1,0 @@
-package gift.service;
-
-public class MemberService {
-}
