+++ src/main/resources/schema.sql
@@ -2,6 +2,7 @@ CREATE TABLE members (
     id BIGINT AUTO_INCREMENT PRIMARY KEY,
     email VARCHAR(320) NOT NULL UNIQUE,
     password VARCHAR(255) NOT NULL
+    role VARCHAR(50) DEFAULT 'USER'
 );
 
 CREATE TABLE products (
